dbt/requirements.txt
anyio==4.1.0
attrs==23.1.0
aws-cdk-lib==2.114.1
aws-cdk.asset-awscli-v1==2.2.201
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
boto3==1.33.11
botocore==1.33.11
cattrs==23.2.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
constructs==10.3.0
contextlib2==21.6.0
cryptography==41.0.7
Deprecated==1.2.14
dill==0.3.7
docker==7.0.0
fastapi==0.95.2
google-pasta==0.2.0
h11==0.14.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
jmespath==1.0.1
jsii==1.93.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
multiprocess==0.70.15
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pathos==0.3.1
platformdirs==4.1.0
pox==0.3.3
ppft==1.7.6.7
protobuf==4.25.1
psutil==5.9.6
publication==0.0.3
pycparser==2.21
pydantic==1.10.13
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.13.2
s3transfer==0.8.2
sagemaker==2.199.0
schema==0.7.5
six==1.16.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.0
starlette==0.27.0
tblib==1.7.0
tqdm==4.66.1
typeguard==2.13.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.22.0
wrapt==1.16.0
zipp==3.17.0
